+++ network/src/index.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import ReactDOM from 'react-dom/client';
-import './index.css';
+import './css/index.css';
 import App from './App';
-import reportWebVitals from './reportWebVitals';
+import reportWebVitals from './js/reportWebVitals';
 import { AppWrapper } from './components/AppWrapper';
 
 const root = ReactDOM.createRoot(document.getElementById('root'));
